# Conflicts:
#	com.unity.render-pipelines.lightweight/Shaders/2D/Include/NormalsRenderingShared.hlsl
#	com.unity.render-pipelines.universal/Editor/ShaderGraph/UniversalPBRSubShader.cs
#	com.unity.render-pipelines.universal/Editor/ShaderGraph/UniversalSpriteLitSubShader.cs
#	com.unity.render-pipelines.universal/Editor/ShaderGraph/UniversalSpriteUnlitSubShader.cs
#	com.unity.render-pipelines.universal/Editor/ShaderGraph/UniversalUnlitSubShader.cs
#	com.unity.render-pipelines.universal/Editor/ShaderGraph/universalUnlitPass.template
#	com.unity.render-pipelines.universal/Shaders/2D/Sprite-Lit-Default.shader
#	com.unity.render-pipelines.universal/Shaders/Particles/ParticlesUnlit.shader
#	com.unity.shadergraph/Editor/Importers/ShaderGraphImporter.cs

+++ com.unity.render-pipelines.universal/Shaders/2D/Include/NormalsRenderingShared.hlsl
@@ -4,7 +4,7 @@
 float4 NormalsRenderingShared(float4 color, float3 normalTS, float3 tangent, float3 bitangent, float3 normal)
 {
     float4 normalColor;
-    float3 normalWS = TransformTangentToWorld(normalTS, half3x3(tangent.xyz, bitangent.xyz, normal.xyz));
+    float3 normalWS = TransformTangentToWorld(normalTS, float3x3(tangent.xyz, bitangent.xyz, normal.xyz));
     float3 normalVS = TransformWorldToViewDir(normalWS);
